cssAttrIf moved to utils.js
