--- src/lib/cssAttrIfHelper.js
@@ -1,2 +1,0 @@
-export const cssAttrIf = (attrKey, attrVal, condition) =>
-	!condition ? '' : `${ attrKey }: ${ attrVal };`;
